+++ spec/integration/resource/api300/c7000/scope/delete_spec.rb
@@ -1,0 +1,27 @@
+# (C) Copyright 2017 Hewlett Packard Enterprise Development LP
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# You may not use this file except in compliance with the License.
+# You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software distributed
+# under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+# CONDITIONS OF ANY KIND, either express or implied. See the License for the
+# specific language governing permissions and limitations under the License.
+
+require 'spec_helper'
+
+klass = OneviewSDK::API300::C7000::Scope
+RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
+  include_context 'integration api300 context'
+
+  describe '#delete' do
+    it 'should delete scope' do
+      items = klass.get_all($client_300)
+      items.each do |item|
+        expect { item.delete }.not_to raise_error
+        expect(item.retrieve!).to eq(false)
+      end
+    end
+  end
+end